scripts/sql-servers-audit-remediation/requirements.txt
azure-identity
azure-mgmt-sql
azure-mgmt-managementgroups
azure-mgmt-policyinsights
azure-mgmt-resource